requirements.txt
asgiref==3.11.0
Django==6.0
django-environ==0.12.0
sqlparse==0.5.5
tzdata==2025.3
